+++ src/lib/abstract-ops/ecmascript.ts
@@ -2,6 +2,16 @@ import { reflectCall } from 'lib/helpers/webidl';
 import { typeIsObject } from '../helpers/miscellaneous';
 import assert from '../../stub/assert';
 
+declare global {
+  interface ArrayBuffer {
+    readonly detached: boolean;
+
+    transfer(): ArrayBuffer;
+  }
+
+  function structuredClone<T>(value: T, options: { transfer: ArrayBuffer[] }): T;
+}
+
 export function CreateArrayFromList<T extends any[]>(elements: T): T {
   // We use arrays to represent lists, so this is basically a no-op.
   // Do a slice though just in case we happen to depend on the unique-ness.
@@ -16,24 +26,33 @@ export function CopyDataBlockBytes(dest: ArrayBuffer,
   new Uint8Array(dest).set(new Uint8Array(src, srcOffset, n), destOffset);
 }
 
-// Not implemented correctly
-export function TransferArrayBuffer<T extends ArrayBufferLike>(O: T): T {
-  return O;
+export let TransferArrayBuffer = (O: ArrayBuffer): ArrayBuffer => {
+  if (typeof O.transfer === 'function') {
+    TransferArrayBuffer = buffer => buffer.transfer();
+  } else if (typeof structuredClone === 'function') {
+    TransferArrayBuffer = buffer => structuredClone(buffer, { transfer: [buffer] });
+  } else {
+    // Not implemented correctly
+    TransferArrayBuffer = buffer => buffer;
+  }
+  return TransferArrayBuffer(O);
+};
+
+export function CanTransferArrayBuffer(O: ArrayBuffer): boolean {
+  return !IsDetachedBuffer(O);
 }
 
-// Not implemented correctly
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function CanTransferArrayBuffer(O: ArrayBufferLike): boolean {
-  return true;
-}
+export let IsDetachedBuffer = (O: ArrayBuffer): boolean => {
+  if (typeof O.detached === 'boolean') {
+    IsDetachedBuffer = buffer => buffer.detached;
+  } else {
+    // Not implemented correctly
+    IsDetachedBuffer = buffer => buffer.byteLength === 0;
+  }
+  return IsDetachedBuffer(O);
+};
 
-// Not implemented correctly
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function IsDetachedBuffer(O: ArrayBufferLike): boolean {
-  return false;
-}
-
-export function ArrayBufferSlice(buffer: ArrayBufferLike, begin: number, end: number): ArrayBufferLike {
+export function ArrayBufferSlice(buffer: ArrayBuffer, begin: number, end: number): ArrayBuffer {
   // ArrayBuffer.prototype.slice is not available on IE10
   // https://www.caniuse.com/mdn-javascript_builtins_arraybuffer_slice
   if (buffer.slice) {